Split TdsParserSafeHandles into separate files Add references to Parser namespace
