+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/PacketHandle.netfx.cs
@@ -6,6 +6,7 @@
 #if NETFRAMEWORK
 
 using System;
+using Microsoft.Data.SqlClient.Parser;
 
 namespace Microsoft.Data.SqlClient
 {
