+++ src/Microsoft.Data.SqlClient/src/Microsoft/Data/SqlClient/SessionHandle.netcore.cs
@@ -5,6 +5,8 @@
 // @TODO: Merge with other implementations (and/or introduce polymorphism to handle this indirection)
 #if NET
 
+using Microsoft.Data.SqlClient.Parser;
+
 namespace Microsoft.Data.SqlClient
 {
     /// <summary>
